rearrange function locations
